Revert the use of Mg::Ptr, back to std::unique_ptr
The idea of this experiment was to check if compile times could be
reduced by avoiding the inclusion of the <memory> header in small
translation units.

However, while #include-ing <memory> expands to about 20 000 lines with
libstdc++, I believe most of those lines are shared between standard
library headers, such that if you include basically any standard header
you get most of that.

The result of this experiment seems to suggest so: compile times did
decrease, but only by about 5%. Hence, it is scarcely worth working
around the standard library.