Make 'RenderCommandData' members reside within RenderCommand struct.
And the rest of the changes are simplifications following from that.

The motivation behind the old design was to hide the actual
RenderCommandData behind a compilation/ABI boundary, so that it could be
arbitrarily different depending on renderer backend. But on reflection,
I see very little that needs to be different depending on backend. The
required adjustments for that can come later.