Fix stack corruption bug in DebugRenderer
This was caused by a misunderstanding of glGetIntegerv. I thought it
always returned wrote single integer (and therefore, the 'v' in the name
befuddled me); now I found out the hard way that how many integers are
written depends on the requested value. GL_POLYGON_MODE will write _two_
values, one for front-face and one for back-face culling mode. This is
fairly useless in modern OpenGL anyway, since it is not allowed to set
those separately anymore anyway.