Remove Mesh-/Texture-/MaterialRepository from GfxDevice
The only reason they were there in the first place is that they hold GPU
state. The GfxDevice was intended to act like a higher-level interface
to the graphics API, and the repositories would be the interface for
uploading data to the GPU.

On further reflection, though, it makes about as much sense to manage
all GPU memory through one class as it would to manage all main-memory
through one class. A bunch of other classes, like the renderers, already
contain GPU state as it is, anyway.

This new design should be more flexible -- it might be very useful to
keep several MeshRepository objects, for example.