Fix compile error
Due to a bad cast that used to work. It is a bit clear what should
happen here. Regardless, looking at the code, the whole OpaqueHandle vs.
OpaqueHandle::Value thing is quite strange and unintuitive (result of
accumulated changes). I the idea is for OpaqueHandle to be owning
(although not RAII-owning) and for OpaqueHandle::Value to be a view. I
should refactor or rename to clarify that intention.