Vertex shader: manually unroll loop to work around IntelHD bug
Got glitchy rendering on animated meshes (apparently random
normal/tangent/bitangent, whenever another dynamic mesh was in view) on
Intel HD 530. All buffers and state appeared correct when inspected in
RenderDoc and apitrace, so it seemed like a driver bug. Manually
unrolling the loop fixed it, so perhaps it was a bug in the GLSL
compiler.