Physics world: clear unused bodies every update
I think this must be how I intended it to work, since the bodies do not
destroy themselves on ref-count 0. It makes sense, since the bodies
would otherwise need to keep track of the world to destroy themselves,
which would probably also be less efficient than doing all the clean up
at once. Perhaps I thought it ought to be the client's responsibility to
call collect_garbage, but that did not work: the physics world must
clear its cache of collisions right before collecting garbage.

I now instead made garbage collection an integrated part of the
physics-world update, and made it clear the collisions cache as part of
that.