Fix issues with coordinate systems in MeshConverter
Turns out I have been overthinking this. Just rotate everything -90
degrees on the X axis to move from AssImp's Y-up to Z-up. Quaternions
should not be affected since they represent rotations, which are
relative. No change of coordinate-system handedness -- right-handed all
along.

The converter could be optimized a bit by precomputing the matrices and
the vector conversion functions but it really isn't a bottleneck for me
at the moment anyway.