Stop trying to separate opengl backend
There was a half-measure attempt at separating OpenGL code into
a minimal subset of files and have a graphics-API-agnostic interface
into the rest of the code, but it was had holes and was not used for
anything anyway. Keeping this separation is too much effort for
a theoretical future extension. If I decide at some point to add Vulkan
support, I'll have to start separating out graphics backends then, not
before.